# docs/OpenAPI.yaml
openapi: 3.0.3
info:
  title: PawPals API
  version: 1.0.0
  description: |
    REST API for the PawPals mobile app. The frontend must not access the database
    or Supabase directly; all data access goes through this API layer.
    Provides CRUD operations for profiles, dogs, matches, playdates, ratings, and messages.
servers:
  - url: https://api.pawpals.example.com/v1
    description: Production server (example)
  - url: http://localhost:4000/v1
    description: Local dev server
...
paths:
...
  /playdates:
    get:
      tags: [Playdates]
      summary: List playdates for current user
      description: Includes playdates where the user owns one of the dogs in the match.
      operationId: listPlaydates
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of playdates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Playdate'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Playdates]
      summary: Create a playdate for a match
      operationId: createPlaydate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaydateCreate'
      responses:
        '201':
          description: Created playdate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playdate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /playdates/{playdateId}:
    get:
      tags: [Playdates]
      summary: Get a playdate by ID
      operationId: getPlaydateById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PlaydateIdParam'
      responses:
        '200':
          description: Playdate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playdate'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Playdate not found
    patch:
      tags: [Playdates]
      summary: Update a playdate
      operationId: updatePlaydate
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PlaydateIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaydateUpdate'
      responses:
        '200':
          description: Updated playdate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playdate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Playdate not found
    delete:
      tags: [Playdates]
      summary: Delete a playdate
      operationId: deletePlaydate
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PlaydateIdParam'
      responses:
        '204':
          description: Playdate deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /matches/{matchId}/messages:
    get:
      tags: [Messages]
      summary: List messages for a match
      operationId: listMessagesForMatch
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MatchIdParam'
      responses:
        '200':
          description: List of messages for this match
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Messages]
      summary: Send a message in a match
      operationId: createMessage
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MatchIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageCreate'
      responses:
        '201':
          description: Created message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /messages/{messageId}:
    get:
      tags: [Messages]
      summary: Get a message by ID
      operationId: getMessageById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MessageIdParam'
      responses:
        '200':
          description: Message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Message not found
    delete:
      tags: [Messages]
      summary: Delete a message
      operationId: deleteMessage
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MessageIdParam'
      responses:
        '204':
          description: Message deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    MatchIdParam:
      in: path
      name: matchId
      required: true
      schema:
        type: string
        format: uuid
    PlaydateIdParam:
      in: path
      name: playdateId
      required: true
      schema:
        type: string
        format: uuid
...
    MessageIdParam:
      in: path
      name: messageId
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized – missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden – user not allowed to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Playdate:
      type: object
      properties:
        playdate_id:
          type: string
          format: uuid
        match_id:
          type: string
          format: uuid
        scheduled_time:
          type: string
          format: date-time
        location:
          type: string
        confirmed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - playdate_id
        - match_id
        - scheduled_time
        - location
        - confirmed
        - created_at
        - updated_at
    PlaydateCreate:
      type: object
      properties:
        match_id:
          type: string
          format: uuid
        scheduled_time:
          type: string
          format: date-time
        location:
          type: string
      required:
        - match_id
        - scheduled_time
        - location
    PlaydateUpdate:
      type: object
      properties:
        scheduled_time:
          type: string
          format: date-time
        location:
          type: string
        confirmed:
          type: boolean
...
    Message:
      type: object
      properties:
        message_id:
          type: string
          format: uuid
        match_id:
          type: string
          format: uuid
        sender_id:
          type: string
          format: uuid
        content:
          type: string
        sent_at:
          type: string
          format: date-time
      required:
        - message_id
        - match_id
        - sender_id
        - content
        - sent_at
    MessageCreate:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    Error:
      type: object
      properties:
        code:
          type: string
          example: BAD_REQUEST
        message:
          type: string
          example: Something went wrong
        details:
          type: object
          additionalProperties: true